Remove node on Port 80 for CryptoCoinGB
Removal of Port 80 for CryptoCoinGB as I was having issues receiving connections on this port, they were very sporadic. As it wasn't doing much use I've decided to take this hardware down temporarily to do some upgrades on it, so would like the port removed from the list for the time being. Thank you :)