Update Node-API-Public-Pools.js
Remove line 7, const JSON = require('circular-json');

It causes 'npm run commands' to fail as it is not a dependency on install and doesn't exist. This has already been removed in the fast loading branch, but is causing issues for anyone completing a fresh install of the terminal miner as it requires the 'circular-json' dependency.

With this line removed, 'npm run commands' works fine and I can see no other effect on the operation of the functions.